+++ src/main/java/com/thebuzzmedia/exiftool/ExifTool.java
@@ -328,7 +328,7 @@ public class ExifTool {
 	 * support one time during each run of the VM and never again during the
 	 * session of that running VM.
 	 */
-	protected static final Map<Feature, Boolean> FEATURE_SUPPORT_MAP = new HashMap<ExifTool.Feature, Boolean>();
+	protected static final Map<Feature, Boolean> FEATURE_SUPPORT_MAP = new HashMap<Feature, Boolean>();
 
 	/**
 	 * Static list of args used to execute ExifTool using the '-ver' flag in
@@ -540,402 +540,20 @@ public class ExifTool {
 		return streams;
 	}
 
-	/**
-	 * Simple class used to house the read/write streams used to communicate
-	 * with an external ExifTool process as well as the logic used to safely
-	 * close the streams when no longer needed.
-	 * <p/>
-	 * This class is just a convenient way to group and manage the read/write
-	 * streams as opposed to making them dangling member variables off of
-	 * ExifTool directly.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	private static class IOStream {
-		BufferedReader reader;
-		OutputStreamWriter writer;
-
-		public IOStream(BufferedReader reader, OutputStreamWriter writer) {
-			this.reader = reader;
-			this.writer = writer;
-		}
-
-		public void close() {
-			try {
-				log("\tClosing Read stream...");
-				reader.close();
-				log("\t\tSuccessful");
-			} catch (Exception e) {
-				// no-op, just try to close it.
-			}
-
-			try {
-				log("\tClosing Write stream...");
-				writer.close();
-				log("\t\tSuccessful");
-			} catch (Exception e) {
-				// no-op, just try to close it.
-			}
-
-			// Null the stream references.
-			reader = null;
-			writer = null;
-
-			log("\tRead/Write streams successfully closed.");
-		}
-	}
-
-	/**
-	 * Enum used to define the different kinds of features in the native
-	 * ExifTool executable that this class can help you take advantage of.
-	 * <p/>
-	 * These flags are different from {@link Tag}s in that a "feature" is
-	 * determined to be a special functionality of the underlying ExifTool
-	 * executable that requires a different code-path in this class to take
-	 * advantage of; for example, <code>-stay_open True</code> support.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Feature {
-		/**
-		 * Enum used to specify that you wish to launch the underlying ExifTool
-		 * process with <code>-stay_open True</code> support turned on that this
-		 * class can then take advantage of.
-		 * <p/>
-		 * Required ExifTool version is <code>8.36</code> or higher.
-		 */
-		STAY_OPEN("8.36");
-
-		/**
-		 * Used to get the version of ExifTool required by this feature in order
-		 * to work.
-		 * 
-		 * @return the version of ExifTool required by this feature in order to
-		 *         work.
-		 */
-		public String getVersion() {
-			return version;
-		}
-
-		private String version;
-
-		private Feature(String version) {
-			this.version = version;
-		}
-	}
-
-	/**
-	 * Enum used to define the 2 different output formats that {@link Tag}
-	 * values can be returned in: numeric or human-readable text.
-	 * <p/>
-	 * ExifTool, via the <code>-n</code> command line arg, is capable of
-	 * returning most values in their raw numeric form (e.g.
-	 * Aperture="2.8010323841") as well as a more human-readable/friendly format
-	 * (e.g. Aperture="2.8").
-	 * <p/>
-	 * While the {@link Tag}s defined on this class do provide a hint at the
-	 * type of the result (see {@link Tag#getType()}), that hint only applies
-	 * when the {@link Format#NUMERIC} form of the value is returned.
-	 * <p/>
-	 * If the caller finds the human-readable format easier to process,
-	 * {@link Format#HUMAN_READABLE} can be specified when calling
-	 * {@link ExifTool#getImageMeta(File, Format, Tag...)} and the returned
-	 * {@link String} values processed manually by the caller.
-	 * <p/>
-	 * In order to see the types of values that are returned when
-	 * {@link Format#HUMAN_READABLE} is used, you can check the comprehensive <a
-	 * href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html">
-	 * ExifTool Tag Guide</a>.
-	 * <p/>
-	 * This makes sense with some values like Aperture that in
-	 * {@link Format#NUMERIC} format end up returning as 14-decimal-place, high
-	 * precision values that are near the intended value (e.g.
-	 * "2.79999992203711" instead of just returning "2.8"). On the other hand,
-	 * other values (like Orientation) are easier to parse when their numeric
-	 * value (1-8) is returned instead of a much longer friendly name (e.g.
-	 * "Mirror horizontal and rotate 270 CW").
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Format {
-		NUMERIC,
-		HUMAN_READABLE;
-	}
-
-	/**
-	 * Enum used to pre-define a convenient list of tags that can be easily
-	 * extracted from images using this class with an external install of
-	 * ExifTool.
-	 * <p/>
-	 * Each tag defined also includes a type hint for the parsed value
-	 * associated with it when the default {@link Format#NUMERIC} value format
-	 * is used.
-	 * <p/>
-	 * All replies from ExifTool are parsed as {@link String}s and using the
-	 * type hint from each {@link Tag} can easily be converted to the correct
-	 * data format by using the provided {@link Tag#parseValue(Tag, String)}
-	 * method.
-	 * <p/>
-	 * This class does not make an attempt at converting the value automatically
-	 * in case the caller decides they would prefer tag values returned in
-	 * {@link Format#HUMAN_READABLE} format and to avoid any compatibility
-	 * issues with future versions of ExifTool if a tag's return value is
-	 * changed. This approach to leaving returned tag values as strings until
-	 * the caller decides they want to parse them is a safer and more robust
-	 * approach.
-	 * <p/>
-	 * The types provided by each tag are merely a hint based on the <a
-	 * href="http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-	 * >ExifTool Tag Guide</a> by Phil Harvey; the caller is free to parse or
-	 * process the returned {@link String} values any way they wish.
-	 * <h3>Tag Support</h3>
-	 * ExifTool is capable of parsing almost every tag known to man (1000+), but
-	 * this class makes an attempt at pre-defining a convenient list of the most
-	 * common tags for use.
-	 * <p/>
-	 * This list was determined by looking at the common metadata tag values
-	 * written to images by popular mobile devices (iPhone, Android) as well as
-	 * cameras like simple point and shoots as well as DSLRs. As an additional
-	 * source of input the list of supported/common EXIF formats that Flickr
-	 * supports was also reviewed to ensure the most common/useful tags were
-	 * being covered here.
-	 * <p/>
-	 * Please email me or <a
-	 * href="https://github.com/thebuzzmedia/imgscalr/issues">file an issue</a>
-	 * if you think this list is missing a commonly used tag that should be
-	 * added to it.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public enum Tag {
-		APERTURE("ApertureValue", Double.class),
-		ARTIST("Artist", String.class),
-		AUTHOR("XPAuthor", String.class),
-		BODY_SERIAL_NUMBER("BodySerialNumber", String.class),
-		CAMERA_SERIAL_NUMBER("CameraSerialNumber", String.class),
-		CAPTION_ABSTRACT("Caption-Abstract", String.class),
-		COLOR_SPACE("ColorSpace", Integer.class),
-		COMMENT("XPComment", String.class),
-		CONTRAST("Contrast", Integer.class),
-		CREATE_DATE("CreateDate", String.class),
-		COPYRIGHT("Copyright", String.class),
-		COPYRIGHT_NOTICE("CopyrightNotice", String.class),
-		CREATION_DATE("CreationDate", String.class),
-		CREATOR("Creator", String.class),
-		DATE_CREATED("DateCreated", String.class),
-		DATE_TIME_ORIGINAL("DateTimeOriginal", String.class),
-		DEVICE_SERIAL_NUMBER("DeviceSerialNumber", String.class),
-		DIGITAL_ZOOM_RATIO("DigitalZoomRatio", Double.class),
-		EXIF_VERSION("ExifVersion", String.class),
-		EXPOSURE_COMPENSATION("ExposureCompensation", Double.class),
-		EXPOSURE_PROGRAM("ExposureProgram", Integer.class),
-		EXPOSURE_TIME("ExposureTime", Double.class),
-		EXTENDER_SERIAL_NUMBER("ExtenderSerialNumber", String.class),
-		FILE_TYPE("FileType", String.class),
-		FLASH("Flash", Integer.class),
-		FLASH_SERIAL_NUMBER("FlashSerialNumber", String.class),
-		FOCAL_LENGTH("FocalLength", Double.class),
-		FOCAL_LENGTH_35MM("FocalLengthIn35mmFormat", Integer.class),
-		GPS_ALTITUDE("GPSAltitude", Double.class),
-		GPS_ALTITUDE_REF("GPSAltitudeRef", Integer.class),
-		GPS_BEARING("GPSDestBearing", Double.class),
-		GPS_BEARING_REF("GPSDestBearingRef", String.class),
-		GPS_DATESTAMP("GPSDateStamp", String.class),
-		GPS_LATITUDE("GPSLatitude", Double.class),
-		GPS_LATITUDE_REF("GPSLatitudeRef", String.class),
-		GPS_LONGITUDE("GPSLongitude", Double.class),
-		GPS_LONGITUDE_REF("GPSLongitudeRef", String.class),
-		GPS_PROCESS_METHOD("GPSProcessingMethod", String.class),
-		GPS_SPEED("GPSSpeed", Double.class),
-		GPS_SPEED_REF("GPSSpeedRef", String.class),
-		GPS_TIMESTAMP("GPSTimeStamp", String.class),
-		IMAGE_HEIGHT("ImageHeight", Integer.class),
-		IMAGE_UNIQUE_ID("ImageUniqueID", String.class),
-		IMAGE_WIDTH("ImageWidth", Integer.class),
-		INTERNAL_SERIAL_NUMBER("InternalSerialNumber", String.class),
-		IPTC_KEYWORDS("Keywords", String.class),
-		ISO("ISO", Integer.class),
-		KEYWORDS("XPKeywords", String.class),
-		LENS("Lens", String.class),
-		LENS_ID("LensID", String.class),
-		LENS_MAKE("LensMake", String.class),
-		LENS_MODEL("LensModel", String.class),
-		LENS_SERIAL_NUMBER("LensSerialNumber", String.class),
-		MAKE("Make", String.class),
-		METERING_MODE("MeteringMode", Integer.class),
-		MODEL("Model", String.class),
-		OBJECT_NAME("ObjectName", String.class),
-		ORIENTATION("Orientation", Integer.class),
-		OWNER_NAME("OwnerName", String.class),
-		RATING("Rating", Integer.class),
-		RATING_PERCENT("RatingPercent", Integer.class),
-		ROTATION("Rotation", Integer.class),
-		SATURATION("Saturation", Integer.class),
-		SCANNER_SERIAL_NUMBER("ScannerSerialNumber", String.class),
-		SENSING_METHOD("SensingMethod", Integer.class),
-		SERIAL_NUMBER("SerialNumber", String.class),
-		SHARPNESS("Sharpness", Integer.class),
-		SHUTTER_SPEED("ShutterSpeedValue", Double.class),
-		SOFTWARE("Software", String.class),
-		SOURCE_SERIAL_NUMBER("SourceSerialNumber", String.class),
-		SUB_SEC_TIME_ORIGINAL("SubSecTimeOriginal", Integer.class),
-		SUBJECT("XPSubject", String.class),
-		TITLE("XPTitle", String.class),
-		WHITE_BALANCE("WhiteBalance", Integer.class),
-		X_RESOLUTION("XResolution", Double.class),
-		Y_RESOLUTION("YResolution", Double.class), ;
-		private static final Map<String, Tag> TAG_LOOKUP_MAP;
-
-		/**
-		 * Initializer used to init the <code>static final</code> tag/name
-		 * lookup map used by all instances of this class.
-		 */
-		static {
-			Tag[] values = Tag.values();
-			TAG_LOOKUP_MAP = new HashMap<String, ExifTool.Tag>(
-					values.length * 3);
-
-			for (int i = 0; i < values.length; i++) {
-				Tag tag = values[i];
-				TAG_LOOKUP_MAP.put(tag.name, tag);
-			}
-		}
-
-		/**
-		 * Used to get the {@link Tag} identified by the given, case-sensitive,
-		 * tag name.
-		 * 
-		 * @param name
-		 *            The case-sensitive name of the tag that will be searched
-		 *            for.
-		 * 
-		 * @return the {@link Tag} identified by the given, case-sensitive, tag
-		 *         name or <code>null</code> if one couldn't be found.
-		 */
-		public static Tag forName(String name) {
-			return TAG_LOOKUP_MAP.get(name);
-		}
-
-		/**
-		 * Convenience method used to convert the given string Tag value
-		 * (returned from the external ExifTool process) into the type described
-		 * by the associated {@link Tag}.
-		 * 
-		 * @param <T>
-		 *            The type of the returned value.
-		 * @param tag
-		 *            The {@link Tag} whose value this is. The tag's type hint
-		 *            will be queried to determine how to convert this string
-		 *            value.
-		 * @param value
-		 *            The {@link String} representation of the tag's value as
-		 *            parsed from the image.
-		 * 
-		 * @return the given string value converted to a native Java type (e.g.
-		 *         Integer, Double, etc.).
-		 * 
-		 * @throws IllegalArgumentException
-		 *             if <code>tag</code> is <code>null</code>.
-		 * @throws NumberFormatException
-		 *             if any exception occurs while trying to parse the given
-		 *             <code>value</code> to any of the supported numeric types
-		 *             in Java via calls to the respective <code>parseXXX</code>
-		 *             methods defined on all the numeric wrapper classes (e.g.
-		 *             {@link Integer#parseInt(String)} ,
-		 *             {@link Double#parseDouble(String)} and so on).
-		 * @throws ClassCastException
-		 *             if the type defined by <code>T</code> is incompatible
-		 *             with the type defined by {@link Tag#getType()} returned
-		 *             by the <code>tag</code> argument passed in. This class
-		 *             performs an implicit/unchecked cast to the type
-		 *             <code>T</code> before returning the parsed result of the
-		 *             type indicated by {@link Tag#getType()}. If the types do
-		 *             not match, a <code>ClassCastException</code> will be
-		 *             generated by the VM.
-		 */
-		@SuppressWarnings("unchecked")
-		public static <T> T parseValue(Tag tag, String value)
-				throws IllegalArgumentException, NumberFormatException {
-			if (tag == null)
-				throw new IllegalArgumentException("tag cannot be null");
-
-			T result = null;
-
-			// Check that there is work to do first.
-			if (value != null) {
-				Class<?> type = tag.type;
-
-				if (Boolean.class.isAssignableFrom(type))
-					result = (T) Boolean.valueOf(value);
-				else if (Byte.class.isAssignableFrom(type))
-					result = (T) Byte.valueOf(Byte.parseByte(value));
-				else if (Integer.class.isAssignableFrom(type))
-					result = (T) Integer.valueOf(Integer.parseInt(value));
-				else if (Short.class.isAssignableFrom(type))
-					result = (T) Short.valueOf(Short.parseShort(value));
-				else if (Long.class.isAssignableFrom(type))
-					result = (T) Long.valueOf(Long.parseLong(value));
-				else if (Float.class.isAssignableFrom(type))
-					result = (T) Float.valueOf(Float.parseFloat(value));
-				else if (Double.class.isAssignableFrom(type))
-					result = (T) Double.valueOf(Double.parseDouble(value));
-				else if (Character.class.isAssignableFrom(type))
-					result = (T) Character.valueOf(value.charAt(0));
-				else if (String.class.isAssignableFrom(type))
-					result = (T) value;
-			}
-
-			return result;
-		}
-
-		/**
-		 * Used to get the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 * 
-		 * @return the name of the tag (e.g. "Orientation", "ISO", etc.).
-		 */
-		public String getName() {
-			return name;
-		}
-
-		/**
-		 * Used to get a hint for the native type of this tag's value as
-		 * specified by Phil Harvey's <a href=
-		 * "http://www.sno.phy.queensu.ca/~phil/exiftool/TagNames/index.html"
-		 * >ExifTool Tag Guide</a>.
-		 * 
-		 * @return a hint for the native type of this tag's value.
-		 */
-		public Class<?> getType() {
-			return type;
-		}
-
-		private String name;
-		private Class<?> type;
-
-		private Tag(String name, Class<?> type) {
-			this.name = name;
-			this.type = type;
-		}
-	}
-
 	private Timer cleanupTimer;
 	private TimerTask currentCleanupTask;
 
-	private IOStream streams;
-	private List<String> args;
+	IOStream streams;
+	List<String> args;
 
-	private Set<Feature> featureSet;
+	Set<Feature> featureSet;
 
 	public ExifTool() {
 		this((Feature[]) null);
 	}
 
 	public ExifTool(Feature... features) throws UnsupportedFeatureException {
-		featureSet = new HashSet<ExifTool.Feature>();
+		featureSet = new HashSet<Feature>();
 
 		if (features != null && features.length > 0) {
 			/*
@@ -1117,8 +735,7 @@ public class ExifTool {
 		 * Create a result map big enough to hold results for each of the tags
 		 * and avoid collisions while inserting.
 		 */
-		Map<Tag, String> resultMap = new HashMap<ExifTool.Tag, String>(
-				tags.length * 3);
+		Map<Tag, String> resultMap = new HashMap<Tag, String>(tags.length * 3);
 
 		if (DEBUG)
 			log("Querying %d tags from image: %s", tags.length,
@@ -1425,7 +1042,7 @@ public class ExifTool {
 	 * It is annoying that we cannot just reset the timer on the task, but that
 	 * isn't the way the java.util.Timer class was designed unfortunately.
 	 */
-	private void resetCleanupTask() {
+	void resetCleanupTask() {
 		// no-op if the timer was never created.
 		if (cleanupTimer == null)
 			return;
@@ -1442,263 +1059,5 @@ public class ExifTool {
 				PROCESS_CLEANUP_DELAY, PROCESS_CLEANUP_DELAY);
 
 		log("\t\tSuccessful");
-	}
-
-	/**
-	 * Class used to represent the {@link TimerTask} used by the internal auto
-	 * cleanup {@link Timer} to call {@link ExifTool#close()} after a specified
-	 * interval of inactivity.
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	private class CleanupTimerTask extends TimerTask {
-		private ExifTool owner;
-
-		public CleanupTimerTask(ExifTool owner) throws IllegalArgumentException {
-			if (owner == null)
-				throw new IllegalArgumentException(
-						"owner cannot be null and must refer to the ExifTool instance creating this task.");
-
-			this.owner = owner;
-		}
-
-		@Override
-		public void run() {
-			log("\tAuto cleanup task running...");
-			owner.close();
-		}
-	}
-
-	/**
-	 * Class used to define an exception that occurs when the caller attempts to
-	 * use a {@link Feature} that the underlying native ExifTool install does
-	 * not support (i.e. the version isn't new enough).
-	 * 
-	 * @author Riyad Kalla (software@thebuzzmedia.com)
-	 * @since 1.1
-	 */
-	public class UnsupportedFeatureException extends RuntimeException {
-		private static final long serialVersionUID = -1332725983656030770L;
-
-		private Feature feature;
-
-		public UnsupportedFeatureException(Feature feature) {
-			super(
-					"Use of feature ["
-							+ feature
-							+ "] requires version "
-							+ feature.version
-							+ " or higher of the native ExifTool program. The version of ExifTool referenced by the system property 'exiftool.path' is not high enough. You can either upgrade the install of ExifTool or avoid using this feature to workaround this exception.");
-		}
-
-		public Feature getFeature() {
-			return feature;
-		}
-	}
-
-	/**
-	 * Class used to handle XMP Sidecar files using exiftool.
-	 * 
-	 * @author Clinton LaForest (clafore@bgsu.edu)
-	 * @since 1.2_thespiritx
-	 */
-	public class Sidecar {
-		public Sidecar(Feature feature) {
-
-		}
-
-		/**
-		 * Used to merge a XMP Sidecar file with an image.
-		 * 
-		 * @param xmp
-		 *            The xmp sidecar file.
-		 * 
-		 * @param file
-		 *            The image file.
-		 * 
-		 * @param preserve
-		 *            <code>true</code> - preserves name mappings
-		 *            <code>false</code> - uses preferred name mappings
-		 * 
-		 * @return <code>void</code>
-		 * 
-		 */
-		public void merge(File xmp, File file, Boolean preserve) {
-			if (xmp == null)
-				throw new IllegalArgumentException(
-						"xmp cannot be null and must be a valid xmp sidecar stream.");
-			if (file == null)
-				throw new IllegalArgumentException("file cannot be null");
-			if (preserve == null)
-				preserve = false;
-			if (!file.canWrite())
-				throw new SecurityException(
-						"Unable to read the given image ["
-								+ file.getAbsolutePath()
-								+ "], ensure that the image exists at the given path and that the executing Java process has permissions to read it.");
-
-			long startTime = System.currentTimeMillis();
-
-			if (DEBUG)
-				log("Writing %s tags to image: %s", xmp.getAbsolutePath(),
-						file.getAbsolutePath());
-
-			long exifToolCallElapsedTime = 0;
-
-			/*
-			 * Using ExifTool in daemon mode (-stay_open True) executes
-			 * different code paths below. So establish the flag for this once
-			 * and it is reused a multitude of times later in this method to
-			 * figure out where to branch to.
-			 */
-			boolean stayOpen = featureSet.contains(Feature.STAY_OPEN);
-
-			args.clear();
-
-			if (stayOpen) {
-				log("\tUsing ExifTool in daemon mode (-stay_open True)...");
-
-				// Always reset the cleanup task.
-				resetCleanupTask();
-
-				/*
-				 * If this is our first time calling getImageMeta with a
-				 * stayOpen connection, set up the persistent process and run it
-				 * so it is ready to receive commands from us.
-				 */
-				if (streams == null) {
-					log("\tStarting daemon ExifTool process and creating read/write streams (this only happens once)...");
-
-					args.add(EXIF_TOOL_PATH);
-					args.add("-stay_open");
-					args.add("True");
-					args.add("-@");
-					args.add("-");
-
-					// Begin the persistent ExifTool process.
-					streams = startExifToolProcess(args);
-				}
-
-				log("\tStreaming arguments to ExifTool process...");
-
-				try {
-					streams.writer.write("-tagsfromfile\n");
-
-					streams.writer.write(file.getAbsolutePath());
-					streams.writer.write("\n");
-
-					if (preserve) {
-						streams.writer.write("-all:all");
-						streams.writer.write("\n");
-					} else {
-						streams.writer.write("-xmp");
-						streams.writer.write("\n");
-					}
-
-					streams.writer.write(xmp.getAbsolutePath());
-					streams.writer.write("\n");
-
-					log("\tExecuting ExifTool...");
-
-					// Begin tracking the duration ExifTool takes to respond.
-					exifToolCallElapsedTime = System.currentTimeMillis();
-
-					// Run ExifTool on our file with all the given arguments.
-					streams.writer.write("-execute\n");
-					streams.writer.flush();
-
-				} catch (IOException e) {
-					log("\tError received in stayopen stream: %s",
-							e.getMessage());
-				} // compact output
-			} else {
-				log("\tUsing ExifTool in non-daemon mode (-stay_open False)...");
-
-				/*
-				 * Since we are not using a stayOpen process, we need to setup
-				 * the execution arguments completely each time.
-				 */
-				args.add(EXIF_TOOL_PATH);
-
-				args.add("-tagsfromfile"); // compact output
-
-				args.add(file.getAbsolutePath());
-
-				if (preserve) {
-					args.add("-all:all");
-				} else {
-					args.add("-xmp");
-				}
-
-				args.add(xmp.getAbsolutePath());
-
-				// Run the ExifTool with our args.
-				streams = startExifToolProcess(args);
-
-				// Begin tracking the duration ExifTool takes to respond.
-				exifToolCallElapsedTime = System.currentTimeMillis();
-			}
-
-			log("\tReading response back from ExifTool...");
-
-			String line = null;
-
-			try {
-				while ((line = streams.reader.readLine()) != null) {
-					/*
-					 * When using a persistent ExifTool process, it terminates
-					 * its output to us with a "{ready}" clause on a new line,
-					 * we need to look for it and break from this loop when we
-					 * see it otherwise this process will hang indefinitely
-					 * blocking on the input stream with no data to read.
-					 */
-					if (stayOpen && line.equals("{ready}"))
-						break;
-				}
-			} catch (IOException e) {
-				log("\tError received in response: %d", e.getMessage());
-			}
-
-			// Print out how long the call to external ExifTool process took.
-			log("\tFinished reading ExifTool response in %d ms.",
-					(System.currentTimeMillis() - exifToolCallElapsedTime));
-
-			/*
-			 * If we are not using a persistent ExifTool process, then after
-			 * running the command above, the process exited in which case we
-			 * need to clean our streams up since it no longer exists. If we
-			 * were using a persistent ExifTool process, leave the streams open
-			 * for future calls.
-			 */
-			if (!stayOpen)
-				streams.close();
-
-			if (DEBUG)
-				log("\tImage Meta Processed in %d ms [write %s tags]",
-						(System.currentTimeMillis() - startTime),
-						xmp.getAbsolutePath());
-
-		}
-
-		/**
-		 * Used to export a XMP Sidecar file from an image.
-		 * 
-		 * @param xmp
-		 *            The xmp sidecar file.
-		 * 
-		 * @param file
-		 *            The image file.
-		 * 
-		 * @param preserve
-		 *            <code>true</code> - preserves name mappings
-		 *            <code>false</code> - uses preferred name mappings
-		 * 
-		 * @return <code>void</code>
-		 * 
-		 */
-		public void export(File xmp, File file, boolean preserve) {
-
-		}
 	}
 }
