+++ src/test/java/com/thebuzzmedia/exiftool/Example.java
@@ -3,10 +3,6 @@ package com.thebuzzmedia.exiftool;
 import java.io.File;
 import java.io.IOException;
 
-import com.thebuzzmedia.exiftool.ExifTool.Feature;
-import com.thebuzzmedia.exiftool.ExifTool.Format;
-import com.thebuzzmedia.exiftool.ExifTool.Tag;
-
 public class Example {
 	public static final int ITERS = 100;
 	public static final File[] IMAGES = new File("src/test/resources")
